+++ Lenguajes/Dotnet/ApuntesNet/ApuntesNet.sln
@@ -222,6 +222,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PluginEjemplo", "Utilities\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Extensions", "Extensions", "{5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AsyncLinq", "Utilities\Extensions\AsyncLinq\AsyncLinq.csproj", "{3A0A883F-834D-48BA-B65A-FB2284715C42}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbConnectionExtensions", "Utilities\Extensions\DbConnectionExtensions\DbConnectionExtensions.csproj", "{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnumExtensions", "Utilities\Extensions\EnumExtensions\EnumExtensions.csproj", "{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging", "Utilities\Extensions\Paging\Paging.csproj", "{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitOfWork", "Utilities\Extensions\UnitOfWork\UnitOfWork.csproj", "{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -508,6 +518,26 @@ Global
 		{96D5971C-1720-40AE-A77F-E241A9C86495}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{96D5971C-1720-40AE-A77F-E241A9C86495}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{96D5971C-1720-40AE-A77F-E241A9C86495}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A0A883F-834D-48BA-B65A-FB2284715C42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A0A883F-834D-48BA-B65A-FB2284715C42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A0A883F-834D-48BA-B65A-FB2284715C42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A0A883F-834D-48BA-B65A-FB2284715C42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -615,6 +645,11 @@ Global
 		{A1FD305E-6CA3-4BBC-B2AC-36D1933E242D} = {6FF492B8-DC9B-4E6E-A8F5-BE4685D006EE}
 		{96D5971C-1720-40AE-A77F-E241A9C86495} = {6FF492B8-DC9B-4E6E-A8F5-BE4685D006EE}
 		{5AC98EFD-A368-4F74-BDBF-2E377C2FBF21} = {0E7AF339-79A8-4190-84EB-B1034CFD0371}
+		{3A0A883F-834D-48BA-B65A-FB2284715C42} = {5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}
+		{B50AA76A-EA14-471F-B1BB-80FAE69DE5EC} = {5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}
+		{EC47205E-9A7A-4488-A5F2-A2FFD46A32AC} = {5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}
+		{3D9CD1DE-D75A-4974-8D44-8B45DBC6BBD4} = {5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}
+		{C26E8ED4-9F8E-4222-A1C1-E6372E75C8AB} = {5AC98EFD-A368-4F74-BDBF-2E377C2FBF21}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E490D0C8-C5A7-4F24-8A9B-CE1635508778}
