+++ Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/DbConnectionExtensions/Extensions/ExecuteCommandExtension.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data;
 using System.Data.Common;
 
-namespace Garciss.DbConnectionExtensions.Extensions
+namespace DbConnectionExtensions.Extensions
 {
     public static class ExecuteCommandExtension
     {
