+++ Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/DbConnectionExtensions/Extensions/QueryAsyncEnumerableExtension.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data;
 using System.Data.Common;
 
-namespace Garciss.DbConnectionExtensions.Extensions;
+namespace DbConnectionExtensions.Extensions;
 
 /// <summary>
 ///
