+++ Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/DbConnectionExtensions/Factory/DbConnectionFactory.cs
@@ -1,6 +1,6 @@
 ﻿using System.Data.Common;
 
-namespace Garciss.DbConnectionExtensions.Factory;
+namespace DbConnectionExtensions.Factory;
 
 internal class DbConnectionFactory : IDbConnectionFactory
 {
