+++ Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/UnitOfWork/Repository/Interfaces/IRepositoryReadOnly.cs
@@ -1,0 +1,6 @@
+﻿namespace UnitOfWork.Repository.Interfaces;
+
+public interface IRepositoryReadOnly<T> : IReadRepository<T> where T : class
+{
+
+}
