--- Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/UnitOfWork/UnitOfWork/DependencyInjection/UnitOfWorkServiceCollectionExtensions.cs
@@ -1,41 +1,0 @@
-﻿using Garciss.UnitOfWork.Repository.Interfaces;
-using Garciss.UnitOfWork.UnitOfWork;
-using Garciss.UnitOfWork.UnitOfWork.Interfaces;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Garciss.UnitOfWork.DependencyInjection;
-
-public static class UnitOfWorkServiceCollectionExtensions
-{
-    public static IServiceCollection AddUnitOfWork<TContext>(this IServiceCollection services)
-       where TContext : DbContext
-    {
-        services.AddScoped<IRepositoryFactory, UnitOfWork<TContext>>();
-        services.AddScoped<IUnitOfWork, UnitOfWork<TContext>>();
-        services.AddScoped<IUnitOfWork<TContext>, UnitOfWork<TContext>>();
-        return services;
-    }
-
-    public static IServiceCollection AddUnitOfWork<TContext1, TContext2>(this IServiceCollection services)
-        where TContext1 : DbContext
-        where TContext2 : DbContext
-    {
-        services.AddScoped<IUnitOfWork<TContext1>, UnitOfWork<TContext1>>();
-        services.AddScoped<IUnitOfWork<TContext2>, UnitOfWork<TContext2>>();
-        return services;
-    }
-
-    public static IServiceCollection AddUnitOfWork<TContext1, TContext2, TContext3>(
-        this IServiceCollection services)
-        where TContext1 : DbContext
-        where TContext2 : DbContext
-        where TContext3 : DbContext
-    {
-        services.AddScoped<IUnitOfWork<TContext1>, UnitOfWork<TContext1>>();
-        services.AddScoped<IUnitOfWork<TContext2>, UnitOfWork<TContext2>>();
-        services.AddScoped<IUnitOfWork<TContext3>, UnitOfWork<TContext3>>();
-
-        return services;
-    }
-}
