--- Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/UnitOfWork/UnitOfWork/Repository/Interfaces/IRepositoryFactory.cs
@@ -1,8 +1,0 @@
-﻿namespace Garciss.UnitOfWork.Repository.Interfaces;
-
-public interface IRepositoryFactory
-{
-    IRepository<T> GetRepository<T>() where T : class;
-    IRepositoryAsync<T> GetRepositoryAsync<T>() where T : class;
-    IRepositoryReadOnly<T> GetReadOnlyRepository<T>() where T : class;
-}
