--- Lenguajes/Dotnet/ApuntesNet/Utilities/Extensions/UnitOfWork/UnitOfWork/Repository/Interfaces/IRepositoryReadOnly.cs
@@ -1,6 +1,0 @@
-﻿namespace Garciss.UnitOfWork.Repository.Interfaces;
-
-public interface IRepositoryReadOnly<T> : IReadRepository<T> where T : class
-{
-
-}
